refactor: move DiffuseWindow class to its own file
